curiosity/requirements.txt
djangorestframework
markdown   
django-filter  
django
PyJWT
requests
jsonschema
